+++ CHANGELOG.md
@@ -6,9 +6,13 @@ In order to read more about upgrading and BC breaks have a look at the [UPGRADE 
 ## 2.3.0
 
 + [#394](https://github.com/luyadev/luya-module-admin/issues/394) Do not run fake cronjob for admin queue if it was not enabled by module's config.
++ [#407](https://github.com/luyadev/luya-module-admin/pull/407) Add new ngrest `raw` plugin which won't change input/output.
++ [#404](https://github.com/luyadev/luya-module-admin/issues/404) NgRest API delete action use ngRestFind() instead of find().
++ [#395](https://github.com/luyadev/luya-module-admin/issues/395) New Active Window to delete tags.
 + [#401](https://github.com/luyadev/luya-module-admin/issues/401) Fixed logout bug for users without file permission.
 + [#397](https://github.com/luyadev/luya-module-admin/issues/397) i18nAttributeFallbackValue() require to run the onFind() context of the given attribute plugin in order to ensure plugin specific options like `markdown`. 
 + [#389](https://github.com/luyadev/luya-module-admin/issues/389) Do not throw an exception by default when pool identifier does not exists in the list of pools.
++ [#403](https://github.com/luyadev/luya-module-admin/pull/403) Use ngRestFind() method for duplicate button instead of find().
 
 ## 2.2.2 (23. October 2019)